refactor: expose config file
