+++ .gitignore
@@ -22,6 +22,5 @@ node_modules
 smart-trader*
 Smart Trader
 SmartTrader*
-env.json
 public
 public*
