+++ config.json
@@ -1,0 +1,3 @@
+{
+  "API_URL":"https://api.binance.com"
+}
