+++ routes/wallet.js
@@ -5,7 +5,7 @@ const fetch = require('node-fetch')
 const Store = require('electron-store')
 const store = new Store()
 const router = express.Router()
-const config = require('../env.json')
+const config = require('../config.json')
 
 // returns all coins balance of a wallet
 router.get('/walletStatus/dump', async (req, res) => {
